MONIT-28: Keep environment in sub-processes.
Monit previously purged the environment so sub-processes started by Monit only had a spartan PATH set and no other environment variables to speak off. The rationale for this functionality was security and control over the path and other env variables. However this created a lot of confusion and problems for users when their programs or script depended on env. vars. and failed because they were not set. It is also a question about of how much security was improved by axing the env. Programs started by Monit do require absolute path etc.