Fix Issue #98: MySQL test might blocked the host because of too many connection errors.
We have to respond with handshake response packet, otherwise the client may be blocked. It seems the restriction is not applied for loopback device and it even wasn't possible to reproduce the issue in lab when testing remote host, but when testing via public interface on the same machine where MySQL runs, then the problem occurred.